requirements.txt
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.9.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.0
frozenlist==1.4.1
future==0.18.3
google-auth==2.40.2
google-genai==1.18.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==3.7.0
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9.1
six==1.16.0
sniffio==1.3.1
typing_extensions==4.13.2
tzdata==2023.4
urllib3==2.1.0
websockets==15.0.1
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
